Instrumentation of generated code with a Ticker
This commit introduces `Ticker`, a new class in `leon.codegen.runtime`,
and updates code generation to count the number of function invocations
during evaluation. The goal is to be able to limit that number, and thus
not have to rely on the JVM throwing a StackOverflowException for
non-terminating executions.

Note that the interface is limited to `CompiledExpression` at this
point, and is not propagated to `Evaluator`. It remains to decide what
constitutes a step exactly, and whether we want the bound on steps to be
part of the default interface for evaluators.